poc for non admin contributors

+++ web/scripts/ms_restore_from_backup_if_necessary.sh
@@ -14,8 +14,16 @@ WIKI_LATEST_TAR_FILE_NAME="mafiascum.backup.$MAFIASCUM_ENVIRONMENT.wiki.latest.z
 if is_mounted_dir_empty "/data/forum" 
 then
     echo "Did not find existing forum volume: restoring from S3 backup..."
-    aws s3 cp "s3://$AWS_BACKUP_BUCKET/web-backups/$FORUM_LATEST_TAR_FILE_NAME" /tmp/$FORUM_LATEST_TAR_FILE_NAME
-    unzip -P${MAFIASCUM_BACKUP_PASSWORD} -d "/data/forum" /tmp/$FORUM_LATEST_TAR_FILE_NAME
+    if [ -z "$AWS_ACCESS_KEY_ID" ]; then
+        aws s3 --no-sign-request cp "s3://$AWS_BACKUP_BUCKET/web-backups/$FORUM_LATEST_TAR_FILE_NAME" /tmp/$FORUM_LATEST_TAR_FILE_NAME
+    else
+        aws s3 cp "s3://$AWS_BACKUP_BUCKET/web-backups/$FORUM_LATEST_TAR_FILE_NAME" /tmp/$FORUM_LATEST_TAR_FILE_NAME
+    fi
+    if [ -z "${MAFIASCUM_BACKUP_PASSWORD}" ]; then
+        unzip -d "/data/forum" /tmp/$FORUM_LATEST_TAR_FILE_NAME
+    else
+        unzip -P${MAFIASCUM_BACKUP_PASSWORD} -d "/data/forum" /tmp/$FORUM_LATEST_TAR_FILE_NAME
+    fi
     rm /tmp/$FORUM_LATEST_TAR_FILE_NAME
     chown -R daemon:root "/data/forum"
 else
@@ -25,9 +33,22 @@ fi
 if is_mounted_dir_empty "/data/wiki" 
 then
     echo "Did not find existing wiki volume: restoring from S3 backup..."
-    aws s3 cp "s3://$AWS_BACKUP_BUCKET/web-backups/$WIKI_LATEST_TAR_FILE_NAME" /tmp/$WIKI_LATEST_TAR_FILE_NAME
-    unzip -P${MAFIASCUM_BACKUP_PASSWORD} -d "/data/wiki" /tmp/$WIKI_LATEST_TAR_FILE_NAME
-    rm /tmp/$WIKI_LATEST_TAR_FILE_NAME
+    if [ -z "$AWS_ACCESS_KEY_ID" ]; then
+        aws s3 --no-sign-request cp "s3://$AWS_BACKUP_BUCKET/web-backups/$WIKI_LATEST_TAR_FILE_NAME" /tmp/$WIKI_LATEST_TAR_FILE_NAME
+    else
+        aws s3 cp "s3://$AWS_BACKUP_BUCKET/web-backups/$WIKI_LATEST_TAR_FILE_NAME" /tmp/$WIKI_LATEST_TAR_FILE_NAME
+    fi
+    if [ $? -eq 0 ]; then
+        if [ -z "${MAFIASCUM_BACKUP_PASSWORD}" ]; then
+            unzip -d "/data/wiki" /tmp/$WIKI_LATEST_TAR_FILE_NAME
+        else
+            unzip -P${MAFIASCUM_BACKUP_PASSWORD} -d "/data/wiki" /tmp/$WIKI_LATEST_TAR_FILE_NAME
+        fi
+        rm /tmp/$WIKI_LATEST_TAR_FILE_NAME
+    else
+        echo "No wiki backup found. Creating empty wiki data directories if needed."
+        mkdir -p /data/wiki/images
+    fi
     chown -R daemon:root "/data/wiki"
 else
     echo "Found existing wiki volume; proceeding."
